Initial visor_wallet tests
